+++ e2e/test-env/src/server/renderer.jsx
@@ -2,24 +2,16 @@ import React from 'react';
 import { renderToString } from 'react-dom/server';
 import RichContentApp from '../../../../examples/main/shared/RichContentApp';
 
-const COMPONENTS = {
-  rce: {
-    Components: [RichContentApp],
-    bundleName: 'index',
-  },
-};
-
 export default function renderer() {
   return (req, res) => {
     const [componentId, fixtureName = 'empty'] = req.path.replace(/^\/|\/$/g, '').split('/');
-    const isMobile = req.query.mobile === '';
-    const locale = req.query.hebrew === '' ? 'he' : 'en';
-    const { Components, bundleName } = COMPONENTS[componentId] || {};
-    const props = { initialState: null, isMobile, locale };
-
-    if (!Components) {
+    if (componentId !== 'rce') {
       return res.status(404).send(`Component for ${componentId} not found`);
     }
+
+    const isMobile = req.query.mobile === '';
+    const locale = req.query.hebrew === '' ? 'he' : 'en';
+    const props = { isMobile, locale };
 
     try {
       props.initialState = require(`../../../tests/fixtures/${fixtureName}.json`);
@@ -29,15 +21,9 @@ export default function renderer() {
     }
 
     res.render('index', {
-      html: renderToString(
-        <>
-          {Components.map((Comp, i) => (
-            <Comp key={i} mode={'test'} {...props} />
-          ))}
-        </>
-      ),
+      html: renderToString(<RichContentApp mode={'test'} {...props} />),
       initialState: props.initialState,
-      bundleName,
+      bundleName: 'index',
       isMobile,
       locale,
     });