+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -125,6 +125,7 @@ const getLinkPanelDropDownConfig = () => {
       searchWords={[searchWords]}
       textToHighlight={textToHighlight}
       highlightTag={({ children }) => <strong className="highlighted-text">{children}</strong>}
+      autoEscape
     />
   );
 
@@ -227,7 +228,7 @@ const videoHandlers = {
   },
 };
 
-export const config = {
+const config = {
   [EMOJI_TYPE]: {
     // toolbar: {
     //   icons: {
@@ -246,8 +247,6 @@ export const config = {
   },
 
   [GALLERY_TYPE]: {
-    scrollingElement: () =>
-      typeof window !== 'undefined' && document.getElementsByClassName('editor-example')[0],
     // toolbar: {
     //   icons: {
     //     InsertPluginButtonIcon: MyCustomIcon,
@@ -662,4 +661,13 @@ export const config = {
     //   })
     // }
   ],
+};
+
+export const getConfig = (additionalConfig = {}) => {
+  let _config = { ...config };
+  Object.keys(additionalConfig).forEach(key => {
+    _config[key] = { ...(_config[key] || {}), ...(additionalConfig[key] || {}) };
+  });
+
+  return _config;
 };