+++ examples/storybook/stories/TestCase/ImageFloatSpacing.js
@@ -3,7 +3,7 @@ import { RichContentEditor, convertFromRaw, createWithContent } from 'wix-rich-c
 import { RichContentViewer } from 'wix-rich-content-viewer';
 import { createImagePlugin } from 'wix-rich-content-plugin-image';
 import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer';
-import imageFloatWithSpace from '../../fixtures/imageFloatWithSpace';
+import imageFloatWithSpace from '../../../../e2e/tests/fixtures/image-float-with-spacing.json';
 import {
   RichContentEditorBox,
   RichContentViewerBox,