+++ examples/storybook/stories/TestCase/index.js
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/react';
-import ImagesStory from './Images.js';
+
 import HTMLPluginStory from './HtmlPlugin.js';
 import BlogLefties from './BlogLefties.js';
 import ImageFloatSpacing from './ImageFloatSpacing.js';
@@ -7,5 +7,4 @@ import ImageFloatSpacing from './ImageFloatSpacing.js';
 storiesOf('Test Cases')
   .add('Image Float Spacing', ImageFloatSpacing)
   .add('Blog Lefties', BlogLefties)
-  .add('HTML Instagram Height', HTMLPluginStory)
-  .add('Images', ImagesStory);
+  .add('HTML Instagram Height', HTMLPluginStory);