+++ packages/plugin-button/web/src/components/button-viewer.jsx
@@ -1,10 +1,10 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { Context, mergeStyles } from 'wix-rich-content-common';
+import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/default-styles.scss';
 class ButtonViewer extends PureComponent {
   render() {
-    const { theme } = this.context || this.props;
+    const { theme } = this.props;
     this.styles = this.styles || mergeStyles({ styles, theme });
     const { url, style, target, rel, buttonText } = this.props;
     return (
@@ -29,8 +29,7 @@ ButtonViewer.propTypes = {
   target: PropTypes.string,
   rel: PropTypes.string,
   buttonText: PropTypes.string,
+  theme: PropTypes.object.isRequired,
 };
-
-ButtonViewer.contextType = Context.type;
 
 export default ButtonViewer;