+++ packages/plugin-html/web/src/index.js
@@ -1,2 +1,3 @@
 export { createHtmlPlugin } from './createHtmlPlugin';
 export { HTML_TYPE } from './types';
+export { pluginHtml } from './editor';