+++ packages/plugin-image/web/src/InPluginInput.jsx
@@ -1,16 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Context } from 'wix-rich-content-common';
 import classnames from 'classnames';
 import styles from '../statics/styles/in-plugin-input.scss';
 
 class InPluginInput extends Component {
   handleFocus = () => {
     this.props.setFocusToBlock();
-    this.context.setInPluginEditingMode(true);
+    this.props.setInPluginEditingMode(true);
   };
 
-  handleBlur = () => this.context.setInPluginEditingMode(false);
+  handleBlur = () => this.props.setInPluginEditingMode(false);
 
   handleKeyPress = e => {
     const { setFocusToBlock, value } = this.props;
@@ -25,12 +24,10 @@ class InPluginInput extends Component {
   className = classnames(styles.inPluginInput, this.props.className);
 
   render() {
-    const { onChange, value } = this.props;
     return (
       <input
         className={this.className}
-        value={value}
-        readOnly={!onChange}
+        value={this.props.value}
         onChange={this.onChange}
         onFocus={this.handleFocus}
         onBlur={this.handleBlur}
@@ -46,8 +43,12 @@ InPluginInput.propTypes = {
   value: PropTypes.string,
   onChange: PropTypes.func,
   setFocusToBlock: PropTypes.func,
+  setInPluginEditingMode: PropTypes.func,
 };
 
-InPluginInput.contextType = Context.type;
+InPluginInput.defaultProps = {
+  setInPluginEditingMode: () => false,
+  setFocusToBlock: () => false,
+};
 
 export default InPluginInput;