+++ packages/plugin-image/web/src/image-component.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import ImageViewer from './image-viewer';
 import { DEFAULTS } from './consts';
 import { sizeClassName, alignmentClassName } from './classNameStrategies';
-import { Context } from 'wix-rich-content-common';
 
 const EMPTY_SMALL_PLACEHOLDER =
   'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7';
@@ -90,7 +89,7 @@ class ImageComponent extends React.Component {
         fileError: null,
       });
     }
-    const { helpers } = this.context;
+    const { helpers } = this.props;
     const hasFileChangeHelper = helpers && helpers.onFilesChange;
     if (hasFileChangeHelper && fileList.length > 0) {
       helpers.onFilesChange(fileList[0], ({ data, error }) =>
@@ -142,9 +141,28 @@ class ImageComponent extends React.Component {
   handleCaptionChange = caption => this.handleMetadataChange({ caption });
 
   render() {
-    const { settings, componentData, onClick, className, blockProps } = this.props;
+    const {
+      settings,
+      componentData,
+      onClick,
+      className,
+      blockProps,
+      theme,
+      isMobile,
+      helpers,
+      disableRightClick,
+      getInPluginEditingMode,
+      setInPluginEditingMode,
+      setComponentUrl,
+    } = this.props;
     return (
       <ImageViewer
+        theme={theme}
+        isMobile={isMobile}
+        helpers={helpers}
+        disableRightClick={disableRightClick}
+        getInPluginEditingMode={getInPluginEditingMode}
+        setInPluginEditingMode={setInPluginEditingMode}
         componentData={componentData}
         onClick={onClick}
         className={className}
@@ -152,15 +170,14 @@ class ImageComponent extends React.Component {
         dataUrl={this.state.dataUrl}
         isFocused={blockProps.isFocused}
         settings={settings}
-        defaultCaption={this.context.t('ImageViewer_Caption')}
+        defaultCaption={this.props.t('ImageViewer_Caption')}
         onCaptionChange={this.handleCaptionChange}
         setFocusToBlock={blockProps.setFocusToBlock}
+        setComponentUrl={setComponentUrl}
       />
     );
   }
 }
-
-ImageComponent.contextType = Context.type;
 
 ImageComponent.propTypes = {
   componentData: PropTypes.object.isRequired,
@@ -171,6 +188,14 @@ ImageComponent.propTypes = {
   onClick: PropTypes.func.isRequired,
   className: PropTypes.string.isRequired,
   settings: PropTypes.object,
+  helpers: PropTypes.object.isRequired,
+  t: PropTypes.func.isRequired,
+  theme: PropTypes.object.isRequired,
+  disableRightClick: PropTypes.bool,
+  getInPluginEditingMode: PropTypes.func,
+  setInPluginEditingMode: PropTypes.func,
+  isMobile: PropTypes.bool.isRequired,
+  setComponentUrl: PropTypes.func,
 };
 
 export { ImageComponent as Component, DEFAULTS };