+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { get, includes, isEqual, isFunction } from 'lodash';
 import {
-  Context,
   mergeStyles,
   validate,
   isSSR,
@@ -41,7 +40,7 @@ class ImageViewer extends React.Component {
   }
 
   getImageUrl(src) {
-    const { helpers, shouldRenderOptimizedImages } = this.context || {};
+    const { helpers, shouldRenderOptimizedImages } = this.props || {};
     if (!src && helpers?.handleFileSelection) {
       return null;
     }
@@ -62,7 +61,7 @@ class ImageViewer extends React.Component {
       } else if (this.state.container) {
         const { width } = this.state.container.getBoundingClientRect();
         requiredWidth = width || src?.width || 1;
-        if (this.context.isMobile) {
+        if (this.props.isMobile) {
           //adjust the image width to viewport scaling and device pixel ratio
           requiredWidth *= (!isSSR() && window.devicePixelRatio) || 1;
           requiredWidth *= (!isSSR() && window.screen.width / document.body.clientWidth) || 1;
@@ -145,7 +144,7 @@ class ImageViewer extends React.Component {
     }
     return (
       <div className={this.styles.imageOverlay}>
-        <Loader type={'medium'} />
+        <Loader type={'medium'} theme={this.props.theme} />
       </div>
     );
   }
@@ -171,14 +170,17 @@ class ImageViewer extends React.Component {
   }
 
   renderCaption(caption) {
-    const { onCaptionChange, setFocusToBlock } = this.props;
-    return (
+    const { onCaptionChange, setFocusToBlock, setInPluginEditingMode } = this.props;
+    return onCaptionChange ? (
       <InPluginInput
+        setInPluginEditingMode={setInPluginEditingMode}
         className={this.styles.imageCaption}
         value={caption}
         onChange={onCaptionChange}
         setFocusToBlock={setFocusToBlock}
       />
+    ) : (
+      <span className={this.styles.imageCaption}>{caption}</span>
     );
   }
 
@@ -195,9 +197,8 @@ class ImageViewer extends React.Component {
   };
 
   shouldRenderCaption() {
-    const { settings, componentData, defaultCaption } = this.props;
+    const { getInPluginEditingMode, settings, componentData, defaultCaption } = this.props;
     const caption = componentData.metadata?.caption;
-    const { getInPluginEditingMode } = this.context;
 
     if (includes(get(settings, 'toolbar.hidden'), 'settings')) {
       return false;
@@ -218,21 +219,21 @@ class ImageViewer extends React.Component {
 
   handleExpand = e => {
     e.preventDefault();
-    const { onExpand } = this.context.helpers;
+    const { onExpand } = this.props.helpers;
     onExpand && onExpand(this.props.entityIndex);
   };
 
-  handleContextMenu = e => this.context.disableRightClick && e.preventDefault();
+  handleContextMenu = e => this.props.disableRightClick && e.preventDefault();
 
   render() {
-    this.styles = this.styles || mergeStyles({ styles, theme: this.context.theme });
-    const { componentData, className, settings } = this.props;
+    this.styles = this.styles || mergeStyles({ styles, theme: this.props.theme });
+    const { componentData, className, settings, setComponentUrl } = this.props;
     const { fallbackImageSrc, ssrDone } = this.state;
     const data = componentData || DEFAULTS;
     const { metadata = {} } = componentData;
 
     const hasLink = data.config && data.config.link;
-    const hasExpand = this.context.helpers && this.context.helpers.onExpand;
+    const hasExpand = this.props.helpers && this.props.helpers.onExpand;
 
     const itemClassName = classNames(this.styles.imageContainer, className, {
       [this.styles.pointer]: hasExpand,
@@ -245,7 +246,8 @@ class ImageViewer extends React.Component {
         ? settings.imageProps(data.src)
         : settings.imageProps;
     }
-    const isGif = imageSrc?.highres?.endsWith('.gif');
+    const isGif = imageSrc?.highres?.endsWith?.('.gif');
+    setComponentUrl?.(imageSrc?.highres);
     const shouldRenderPreloadImage = imageSrc && !isGif;
     const shouldRenderImage = (imageSrc && ssrDone) || isGif;
 
@@ -278,8 +280,6 @@ class ImageViewer extends React.Component {
   }
 }
 
-ImageViewer.contextType = Context.type;
-
 ImageViewer.propTypes = {
   componentData: PropTypes.object.isRequired,
   className: PropTypes.string,
@@ -291,6 +291,13 @@ ImageViewer.propTypes = {
   entityIndex: PropTypes.number,
   onCaptionChange: PropTypes.func,
   setFocusToBlock: PropTypes.func,
+  theme: PropTypes.object.isRequired,
+  helpers: PropTypes.object.isRequired,
+  disableRightClick: PropTypes.bool,
+  getInPluginEditingMode: PropTypes.func,
+  setInPluginEditingMode: PropTypes.func,
+  isMobile: PropTypes.bool.isRequired,
+  setComponentUrl: PropTypes.func,
 };
 
 export default ImageViewer;