+++ packages/viewer/web/src/RichContentViewer.jsx
@@ -5,7 +5,6 @@ import {
   mergeStyles,
   AccessibilityListener,
   normalizeInitialState,
-  Context,
   getLangDir,
 } from 'wix-rich-content-common';
 import { convertToReact } from './utils/convertContentState';
@@ -69,35 +68,47 @@ class RichContentViewer extends Component {
     };
   }
 
+  static getDerivedStateFromError(error) {
+    return { error };
+  }
+
   render() {
-    const { styles } = this;
-    const { textDirection, typeMappers, decorators, inlineStyleMappers, locale } = this.props;
+    const { onError } = this.props;
+    try {
+      if (this.state.error) {
+        onError(this.state.error);
+        return null;
+      }
+      const { styles } = this;
+      const { textDirection, typeMappers, decorators, inlineStyleMappers, locale } = this.props;
 
-    const wrapperClassName = classNames(styles.wrapper, {
-      [styles.desktop]: !this.props.platform || this.props.platform === 'desktop',
-    });
-    const editorClassName = classNames(styles.editor, {
-      [styles.rtl]: textDirection === 'rtl',
-    });
+      const wrapperClassName = classNames(styles.wrapper, {
+        [styles.desktop]: !this.props.platform || this.props.platform === 'desktop',
+      });
+      const editorClassName = classNames(styles.editor, {
+        [styles.rtl]: textDirection === 'rtl',
+      });
 
-    const output = convertToReact(
-      this.state.raw,
-      styles,
-      textDirection,
-      typeMappers,
-      this.state.contextualData,
-      decorators,
-      inlineStyleMappers
-    );
+      const output = convertToReact(
+        this.state.raw,
+        styles,
+        textDirection,
+        typeMappers,
+        this.state.contextualData,
+        decorators,
+        inlineStyleMappers
+      );
 
-    return (
-      <div className={wrapperClassName} dir={getLangDir(locale)}>
-        <Context.Provider value={this.state.contextualData}>
+      return (
+        <div className={wrapperClassName} dir={getLangDir(locale)}>
           <div className={editorClassName}>{output}</div>
-          <AccessibilityListener />
-        </Context.Provider>
-      </div>
-    );
+          <AccessibilityListener isMobile={this.props.isMobile} />
+        </div>
+      );
+    } catch (err) {
+      onError(err);
+      return null;
+    }
   }
 }
 
@@ -131,6 +142,7 @@ RichContentViewer.propTypes = {
   disabled: PropTypes.bool,
   shouldRenderOptimizedImages: PropTypes.bool,
   siteDomain: PropTypes.string,
+  onError: PropTypes.func,
 };
 
 RichContentViewer.defaultProps = {
@@ -139,6 +151,9 @@ RichContentViewer.defaultProps = {
   typeMappers: [],
   inlineStyleMappers: [],
   locale: 'en',
+  onError: err => {
+    throw err;
+  },
 };
 
 export default RichContentViewer;