+++ packages/wrapper/web/src/PluginsStrategy/index.js
@@ -1,0 +1,2 @@
+import pluginsStrategyProvider from './PluginsStrategyProvider';
+export default pluginsStrategyProvider;